Refactor user profile and chatbot AI components
- Removed Points, Promotions, Reviews, Rewards, and other related pages from the user profile layout.
- Introduced new layout structure for vendor detail and search functionalities.
- Added new pages for vendor FAQs, packages, portfolio, reviews, and about sections.
- Implemented dynamic layouts for chatbot AI and guest interactions.
- Enhanced user session management and data fetching for various components.
- Cleaned up unused imports and optimized component structures.
