+++ src/app/(user)/(chatbotAI)/(guest)/(vendor)/[slug]/faq/page.tsx
