+++ src/app/(user)/(chatbotAI)/(guest)/(vendor)/[slug]/layout.tsx
