+++ src/app/(user)/(chatbotAI)/(guest)/(vendor)/[slug]/packages/page.tsx
