+++ src/app/(user)/(chatbotAI)/(guest)/(vendor)/[slug]/page.tsx
