+++ src/app/(user)/(chatbotAI)/(guest)/(vendor)/[slug]/portfolio/page.tsx
