+++ src/app/(user)/(chatbotAI)/(guest)/layout.tsx
@@ -6,21 +6,21 @@ import userService from "@services/user";
 import { IUser } from "@models/user/common.model";
 import { IUserResponse } from "@models/user/response.model";
 import Footer from "@components/Organisms/Footer";
-import HeaderHomePage from "@components/Organisms/HeaderHomePage";
-import packageServices from "@services/packageServices";
+import Header from "@components/Organisms/Header";
 import { IServicePackagesData } from "@models/servicePackages/response.model";
-import { METADATA } from "../../../types/IMetadata";
+import packageServices from "@services/packageServices";
+import { METADATA } from "../../../../types/IMetadata";
 import Chatbot from "@components/Molecules/Chatbot";
 
 async function getAUser(id: string) {
     return await userService.getAUser(id);
 }
 
-async function getAllServicePackage(current: number = 1, pageSize: number = 10, showAll: boolean = true, sortBy: string = "createdAt", sortDirection: string = "asc") {
-    return await packageServices.getAllServicePackage(current, pageSize, showAll, sortBy, sortDirection);
+async function getAllServicePackage() {
+    return await packageServices.getAllServicePackage();
 }
 
-export default async function RootLayout({
+export default async function GuestLayout({
     children,
 }: Readonly<{
     children: React.ReactNode;
@@ -31,13 +31,12 @@ export default async function RootLayout({
         const user = await getAUser(session.user.id) as IUserResponse;
         userData = user?.data as IUser | undefined;
     }
-    const servicePackages = await getAllServicePackage(1, 1000, true, "createdAt", "asc");
+    const servicePackages = await getAllServicePackage();
     const servicePackagesData = servicePackages?.data as IServicePackagesData;
     return (
         <>
-            <HeaderHomePage user={userData} servicePackages={servicePackagesData} />
+            <Header user={userData} servicePackages={servicePackagesData} />
             {children}
-            <Chatbot />
             <Footer />
         </>
     );
