+++ src/app/(user)/(chatbotAI)/(guest)/search/packages/page.tsx
