+++ src/app/(user)/(chatbotAI)/(guest)/search/vendors/page.tsx
